feat(Datasets): DatasetLayout component
