+++ src/pages/workspaces/[workspaceSlug]/datasets/[datasetSlug]/management/index.tsx
@@ -1,17 +1,7 @@
-import {
-  CloudArrowUpIcon,
-  PlusIcon,
-  TrashIcon,
-} from "@heroicons/react/24/outline";
-import Breadcrumbs from "core/components/Breadcrumbs";
 import Button from "core/components/Button/Button";
-import DataCard from "core/components/DataCard";
-import TextProperty from "core/components/DataCard/TextProperty";
 import Page from "core/components/Page";
 import { createGetServerSideProps } from "core/helpers/page";
 import { NextPageWithLayout } from "core/helpers/types";
-import DatasetVersionFilesDataGrid from "datasets/features/DatasetVersionFilesDataGrid/DatasetVersionFilesDataGrid";
-import DatasetVersionPicker from "datasets/features/DatasetVersionPicker/DatasetVersionPicker";
 import { useTranslation } from "next-i18next";
 import { useRouter } from "next/router";
 import {
@@ -20,27 +10,13 @@ import {
   WorkspaceDatasetPageQuery,
   WorkspaceDatasetPageQueryVariables,
 } from "workspaces/graphql/queries.generated";
-import WorkspaceLayout from "workspaces/layouts/WorkspaceLayout";
 import { useEffect, useState } from "react";
-import { updateDataset } from "datasets/helpers/dataset";
-import UploadDatasetVersionDialog from "datasets/features/UploadDatasetVersionDialog";
-import DescriptionList from "core/components/DescriptionList";
-import Time from "core/components/Time";
-import PinDatasetButton from "datasets/features/PinDatasetButton";
-import LinkDatasetDialog from "datasets/features/LinkDatasetDialog";
 import { LinkIcon } from "@heroicons/react/24/solid";
 import useCacheKey from "core/hooks/useCacheKey";
-import DeleteDatasetTrigger from "datasets/features/DeleteDatasetTrigger";
-import RenderProperty from "core/components/DataCard/RenderProperty";
-import Clipboard from "core/components/Clipboard";
 import { trackEvent } from "core/helpers/analytics";
-import Tabs from "core/components/Tabs";
-import Block from "core/components/Block";
-import Title from "core/components/Title";
-import DatasetExplorer from "datasets/features/DatasetExplorer";
-import { capitalize } from "lodash";
 import DatasetTabs from "datasets/features/DatasetTabs/DatasetTabs";
 import DatasetLinksDataGrid from "datasets/features/DatasetLinksDataGrid";
+import DatasetLayout from "datasets/features/layouts/DatasetLayout";
 
 export type WorkspaceDatabasePageProps = {
   datasetSlug: string;
@@ -81,150 +57,33 @@ const WorkspaceDatasetAccessManagementPage: NextPageWithLayout = (
     }
   }, []);
 
-  const onChangeVersion: React.ComponentProps<
-    typeof DatasetVersionPicker
-  >["onChange"] = (version) => {
-    router.push({
-      pathname: router.pathname,
-      query: { ...router.query, version: version?.id },
-    });
-  };
-
   if (!data?.datasetLink) {
     return null;
   }
   const { datasetLink } = data;
   const { dataset, workspace } = datasetLink;
-  const isWorkspaceSource = workspace.slug === dataset.workspace?.slug;
-  const version = dataset.version || dataset.latestVersion || null;
-
-  const onSave = async (values: any) => {
-    await updateDataset(datasetLink.dataset.id, values);
-  };
 
   return (
     <Page title={datasetLink.dataset.name ?? t("Dataset")}>
-      <WorkspaceLayout
-        workspace={workspace}
-        helpLinks={[
-          {
-            label: t("About datasets"),
-            href: "https://github.com/BLSQ/openhexa/wiki/User-manual#datasets",
-          },
-          {
-            label: t("Using the OpenHEXA SDK with datasets"),
-            href: "https://github.com/BLSQ/openhexa/wiki/Using-the-OpenHEXA-SDK#working-with-datasets",
-          },
-        ]}
-      >
-        <WorkspaceLayout.Header className="flex items-center justify-between gap-2">
-          <Breadcrumbs withHome={false} className="flex-1">
-            <Breadcrumbs.Part
-              isFirst
-              href={`/workspaces/${encodeURIComponent(workspace.slug)}`}
-            >
-              {workspace.name}
-            </Breadcrumbs.Part>
-            <Breadcrumbs.Part
-              href={`/workspaces/${encodeURIComponent(
-                workspace.slug,
-              )}/datasets`}
-            >
-              {t("Datasets")}
-            </Breadcrumbs.Part>
-            <Breadcrumbs.Part
-              isLast
-              href={`/workspaces/${encodeURIComponent(
-                workspace.slug,
-              )}/datasets/${encodeURIComponent(datasetLink.id)}`}
-            >
-              {datasetLink.dataset.name}
-            </Breadcrumbs.Part>
-          </Breadcrumbs>
-          <PinDatasetButton link={datasetLink} />
-          {datasetLink.dataset.permissions.createVersion &&
-            isWorkspaceSource && (
-              <Button
-                leadingIcon={<PlusIcon className="h-4 w-4" />}
-                onClick={() => setUploadDialogOpen(true)}
-              >
-                {t("Create new version")}
-              </Button>
-            )}
-          {isWorkspaceSource && datasetLink.dataset.permissions.delete && (
-            <DeleteDatasetTrigger
-              dataset={datasetLink.dataset}
-              onDelete={() =>
-                router.push({
-                  pathname: "/workspaces/[workspaceSlug]/datasets",
-                  query: { workspaceSlug: workspace.slug },
-                })
-              }
-            >
-              {({ onClick }) => (
+      <DatasetLayout datasetLink={data.datasetLink} workspace={workspace}>
+        <DatasetTabs datasetLink={datasetLink} currentTab="management">
+          <>
+            <div className={"flex flex justify-end"}>
+              {workspace.permissions.update && (
                 <Button
-                  variant={"danger"}
-                  onClick={onClick}
-                  leadingIcon={<TrashIcon className="w-4" />}
+                  leadingIcon={<LinkIcon className={"h-4 w-4"} />}
+                  onClick={() => setLinkDialogOpen(true)}
                 >
-                  {t("Delete")}
+                  {t("Share with a workspace")}
                 </Button>
               )}
-            </DeleteDatasetTrigger>
-          )}
-        </WorkspaceLayout.Header>
-        <WorkspaceLayout.PageContent>
-          <Block className="py-4 min-h-[70vh]">
-            <Block.Header className="flex gap-4 items-center justify-between">
-              {capitalize(datasetLink.dataset.name)}
-              {version && (
-                <DatasetVersionPicker
-                  onChange={onChangeVersion}
-                  dataset={datasetLink.dataset}
-                  version={version}
-                  className="w-40"
-                />
-              )}
-            </Block.Header>
-            <Block.Content className="space-y-2">
-              <DatasetTabs
-                datasetSlug={datasetSlug}
-                workspaceSlug={workspaceSlug}
-                currentTab="management"
-                isWorkspaceSource={isWorkspaceSource}
-                version={version}
-              >
-                <>
-                  <div className={"flex flex justify-end"}>
-                    {workspace.permissions.update && (
-                      <Button
-                        leadingIcon={<LinkIcon className={"h-4 w-4"} />}
-                        onClick={() => setLinkDialogOpen(true)}
-                      >
-                        {t("Share with a workspace")}
-                      </Button>
-                    )}
-                  </div>
-                  <div className={"-mx-6"}>
-                    <DatasetLinksDataGrid dataset={datasetLink.dataset} />
-                  </div>
-                </>
-              </DatasetTabs>
-            </Block.Content>
-          </Block>
-        </WorkspaceLayout.PageContent>
-      </WorkspaceLayout>
-
-      <UploadDatasetVersionDialog
-        open={isUploadDialogOpen}
-        onClose={() => setUploadDialogOpen(false)}
-        datasetLink={datasetLink}
-      />
-      <LinkDatasetDialog
-        dataset={datasetLink.dataset}
-        open={isLinkDialogOpen}
-        onClose={() => setLinkDialogOpen(false)}
-      />
+            </div>
+            <div className={"-mx-6"}>
+              <DatasetLinksDataGrid dataset={datasetLink.dataset} />
+            </div>
+          </>
+        </DatasetTabs>
+      </DatasetLayout>
     </Page>
   );
 };
